Add files via upload
Helloooo
It's Day 2 of the Challenge.
We moving from simple to complex.

DAY 2
Write a program which iterates the integers from 1 to 50. For multiples of three print "Fizz" and for multiples of five print "Buzz". For numbers which are multiples of three and five, print "FizzBuzz".

Note
Multiples of three = 3, 6, 9, 12...

Multiples of five  = 5, 15, 20, 25...

Multiples of three and five = 15, 30, ...

What we expect = 1, 2, "Fizz", 4, "Buzz", "Fuzz", 7, 8, "Fizz", "Buzz", ..., 14, "FizzBuzz"...

Submission: on or before 11:00pm.